sql, server: fix periodic resetting of stmt stats
Addressing own comments from #24511: we have a loop resetting some stats
but it's resetting the wrong ones - the ones on the old Executor,
unused. It should be resetting the sql.Server stats.
Also moved the resetting code from server.Server to sql.Server, where
it belongs.

Release note: None
